checkers/buggy/requirements.txt
pwntools

